Simplify internal state (#58)
* simplify the state by removing hasCompleteRow

hasCompleteRow can be computed, no need to persist it in the state.
It's fast because we don't have more than 50-100 rows.

Also: I renamed SET_ROWS into FETCHED_ROWS_SLICE, as conceptually a
resolver takes action and decide what to do with it. But the action as
itself only described an event that occurred, not what the resolver
should do with it. Anyway, it's a NIT

* simplify the state: no need for invalidation

* update the CHANGELOG

* merge two changelog lines

* refactor

* microoptimize a bit

* turn hasCompleteRow to a state

* move the code, to avoid two renders