make the scrollable div focusable
and remove it from the table itself. When focused, the table can be
scrolled horizontally and vertically if needed with the arrows.

The Tab sequence only contains the scrollable element at the moment.

Before, it only contained the table (in Chrome) while it contained two
elements in Firefox (that already added the scrollable div to the
sequence).