Fix the sort arrow icons! Also: add description about sort
With the current setup, the accessible name
(https://developer.mozilla.org/en-US/docs/Glossary/Accessible_name)
is given by the content of the <td> element, and the accessible
description
(https://developer.mozilla.org/en-US/docs/Glossary/Accessible_description)
is given by the aria-description attribute.
We also show the description in the title popup, to help understand the
sort behavior.
Note that the description provides information about the action when
clicking, not about the current state, which is already provided by
the icons and the aria-sort attribute.