fix the sort icons + focus the table + improve title/description of columns (#135)
* make the scrollable div focusable

and remove it from the table itself. When focused, the table can be
scrolled horizontally and vertically if needed with the arrows.

The Tab sequence only contains the scrollable element at the moment.

Before, it only contained the table (in Chrome) while it contained two
elements in Firefox (that already added the scrollable div to the
sequence).

* add aria label (using caption element) to give context when focusing

See https://inclusive-components.design/data-tables/. We use caption +
aria-labelledby, instead of aria-label, because it can be translated.

* Fix the sort arrow icons! Also: add description about sort

With the current setup, the accessible name
(https://developer.mozilla.org/en-US/docs/Glossary/Accessible_name)
is given by the content of the <td> element, and the accessible
description
(https://developer.mozilla.org/en-US/docs/Glossary/Accessible_description)
is given by the aria-description attribute.
We also show the description in the title popup, to help understand the
sort behavior.
Note that the description provides information about the action when
clicking, not about the current state, which is already provided by
the icons and the aria-sort attribute.

* memoize the description

* show black outline (with border, not outline) when table is focused

it's the same as what Chrome shows by default, but now it's explicit and
cross browser

* no need to add tabindex (following my tests) + an element with tabindex should be interactive, and scolling is not interaction