Use data attributes, and add aria-colindex (#139)
* add aria-readonly to all cells, as for now hightable does not let edit the content

* use custom data- attributes instead of hacking aria- attributes

Instead of `aria-posinset` and `aria-setsize`, use
`data-order-by-index` and `data-order-by-size`, to avoid misleading
the users. Multicolumn sort is not supported by ARIA at the moment,
so didn't try to be clever. See https://github.com/w3c/aria/issues/283

* explicitly set aria-colindex

* Revert "add aria-readonly to all cells, as for now hightable does not let edit the content"

This reverts commit dc312f9a715bd5e215cfa978ecd1aedc38e49566.

It's not needed (and even recommended against, see
https://www.w3.org/WAI/ARIA/apg/patterns/grid/#wai-ariaroles,states,andproperties)
because aria-readonly is already set on the "grid" role element.

* add comment